UserGuidanceAI/requirements.txt
albumentations==1.3.0
comet_ml==3.32.4
coremltools==6.2
Flask==2.2.3
matplotlib==3.7.0
mss==7.0.1
onnxruntime==1.14.1
onnxsim==0.4.17
opencv_python==4.7.0.72
openvino==2022.3.0
paddle==1.0.2
pafy==0.5.5
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pycocotools==2.0.6
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
setuptools==58.1.0
tensorflow==2.11.0
tensorflowjs==4.2.0
tensorrt==8.6.0
tflite_runtime==2.11.0
tflite_support==0.4.3
thop==0.1.1.post2209072238
torchvision==0.15.1+cu117
tqdm==4.64.1
tritonclient==2.31.0
x2paddle==1.4.0
